src/app/gateway/requirements.txt
aiokafka==0.12.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
certifi==2025.11.12
cffi==2.0.0
click==8.3.0
confluent-kafka==2.12.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.119.0
fluent-logger==0.11.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jose==1.0.0
jwt==1.4.0
kafka==1.3.5
logger==1.4
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==25.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.0
pydantic-settings==2.12.0
pydantic_core==2.41.1
python-dotenv==1.1.1
python-http-client==3.3.7
python-jose==3.5.0
rsa==4.9.1
sendgrid==6.12.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
Werkzeug==3.1.3
